Control.Concurrent.FreshVar

Description

A FreshVar is a shared readable value that always remains fresh. One can be created and read simply by providing a stale check and a create/refresh action.

>>> var <- newFreshVar alwaysFresh onRefreshFalse
>>> readFreshVar var
True

When a value loses freshness the create/refresh action is run to refresh it.

>>> var <- newFreshVar alwaysStale onRefreshFalse
>>> readFreshVar var
False

Refreshes can also be preemptively triggered by providing an early warning check. These refreshes happen async and can prevent excessive blocking when values become stale.

>>> import Control.Concurrent (threadDelay)
>>> 
>>> var <- newPreemptiveFreshVar alwaysFresh alwaysStale onRefreshFalse
>>> threadDelay 1
>>> readFreshVar var
False

Documentation

data FreshVar a

A value that is always fresh

newFreshVar

Arguments

:: (a -> Bool)
A check to determine if the value is stale

-> (Maybe a -> IO a)
A procedure to create or refresh the value

-> IO (FreshVar a)

Create a value that will always remain fresh

A FreshVar will refresh itself when its stale check returns True. These refreshes are done lazily and occur when a stale value is read via readFreshVar.

newPreemptiveFreshVar

Arguments

:: (a -> Bool)
A check to determine if the value is stale

-> (a -> Bool)
A check to determine if that value is nearing stale

-> (Maybe a -> IO a)
A procedure to create or refresh the value

-> IO (FreshVar a)

Create a FreshVar that preemptively refreshes itself

A FreshVar will block reads when the value becomes stale. However a preemptive FreshVar can refresh itself before the value becomes stale and prevent blocking reads.

readFreshVar :: Show a => FreshVar a -> IO a

Read a value and ensure it is always fresh